AOCDay7
Starting code for Day 7, probably need to read up on Regex, 
For directory, the plan is to use structs to keep track of path directory, 
But could Also use stack structure. Also with control flow, need to be able to take the ls command
read in lines until the next $ encountered. As well as taking in ints as sizes. Also need to handle duplicate directory
edge cases