Day 7 Solution Part 2
Ended up solving recursively. Had issues with trying to parse in the $ for commands... but I just managed to figure it out and just wild card the character to actual be able to capture it. The general logic of parsing the directory tree was daunting when I first encountered it, but after thinking about it for a while, it came pretty easily, and I was able to type out the logic pretty easily. Had to use that rogue magic number as a sentinel though... Recursively calculated the directories that would free up the capcacity. Not really all that elegant of a solution and a total mess to look at.